+++ src/pages/NotificationsRedesign/redesign/ui/index.js
@@ -1,0 +1,1 @@
+export * from './ui';
