+++ src/providers/Notifications.js
@@ -204,7 +204,6 @@ class NotificationsProvider extends React.Component {
   }
 
   processNotificationsChunk = (nextPage, notificationsChunk) => {
-    console.warn('notificationsChunk', notificationsChunk)
     return new Promise((resolve, reject) => {
       let everythingUpdated = true;
